Move manufacturerId from basic to options command and manage unassignement
