+++ src/Core/Domain/Product/Command/UpdateProductBasicInformationCommand.php
@@ -28,7 +28,6 @@ declare(strict_types=1);
 
 namespace PrestaShop\PrestaShop\Core\Domain\Product\Command;
 
-use PrestaShop\PrestaShop\Core\Domain\Manufacturer\ValueObject\ManufacturerId;
 use PrestaShop\PrestaShop\Core\Domain\Product\ValueObject\ProductId;
 
 /**
@@ -60,11 +59,6 @@ class UpdateProductBasicInformationCommand
      * @var string[]|null key value pairs where key is the id of language
      */
     private $localizedShortDescriptions;
-
-    /**
-     * @var ManufacturerId|null
-     */
-    private $manufacturerId;
 
     /**
      * @param int $productId
@@ -158,26 +152,6 @@ class UpdateProductBasicInformationCommand
     public function setLocalizedShortDescriptions(array $localizedShortDescriptions): UpdateProductBasicInformationCommand
     {
         $this->localizedShortDescriptions = $localizedShortDescriptions;
-
-        return $this;
-    }
-
-    /**
-     * @return ManufacturerId|null
-     */
-    public function getManufacturerId(): ?ManufacturerId
-    {
-        return $this->manufacturerId;
-    }
-
-    /**
-     * @param int $manufacturerId
-     *
-     * @return UpdateProductBasicInformationCommand
-     */
-    public function setManufacturerId(int $manufacturerId): UpdateProductBasicInformationCommand
-    {
-        $this->manufacturerId = new ManufacturerId($manufacturerId);
 
         return $this;
     }
